refactor: consolidate DDNS profile form validation and enable IPv4/IPv6 settings + bugfix broken ipv4 and ipv6 parameter + fix url display to new form
