+++ routes/(authenticated)/ddns-profiles/[id]/edit.tsx
@@ -9,43 +9,7 @@ import { db } from "@/lib/db/index.ts";
 import { DDNSProfilesTable } from "@/lib/db/schemas/ddns_profiles.table.ts";
 import { ConnectedServicesTable } from "@/lib/db/schemas/connected_services.table.ts";
 import { and, eq } from "drizzle-orm";
-
-const DnsRecordSchema = z.object({
-  record_name: z.string().min(1),
-  zone_id: z.string().min(1),
-});
-
-const CheckboxSchema = z.literal("on").optional().transform((value) =>
-  value === "on"
-);
-
-const EditDDNSProfileSchema = z.object({
-  profile_name: z.string().min(1),
-  connected_service_id: z.uuid(),
-  dns_records: z.string().transform((str, ctx) => {
-    try {
-      const parsed = JSON.parse(str);
-      const result = z.array(DnsRecordSchema).safeParse(parsed);
-      if (!result.success) {
-        ctx.addIssue({
-          code: "custom",
-          message: "Invalid DNS records format",
-        });
-        return z.NEVER;
-      }
-      return result.data;
-    } catch {
-      ctx.addIssue({
-        code: "custom",
-        message: "DNS records must be valid JSON",
-      });
-      return z.NEVER;
-    }
-  }),
-  allowed_user_agent: z.string().optional(),
-  ipv4_enabled: CheckboxSchema,
-  ipv6_enabled: CheckboxSchema,
-});
+import { DDNSProfileFormSchema } from "@/schemas/ddns/profileForm.ts";
 
 export default define.page(async (ctx) => {
   const authResult = await ctx.state.authPromise;
@@ -88,7 +52,7 @@ export default define.page(async (ctx) => {
   // HANDLE POST REQUEST
   if (ctx.req.method === "POST") {
     const formData = await ctx.req.formData();
-    const parsedInput = EditDDNSProfileSchema.safeParse(
+    const parsedInput = DDNSProfileFormSchema.safeParse(
       Object.fromEntries(formData),
     );
 
@@ -276,7 +240,7 @@ export default define.page(async (ctx) => {
       <div class="p-4 bg-blue-50 border border-blue-300 rounded text-sm">
         <h4 class="font-bold mb-2">ℹ️ Current DDNS Endpoint:</h4>
         <code class="text-sm break-all">
-          POST {ctx.url.origin}/ddns/{profile.id}?ip=YOUR_IP
+          POST {ctx.url.origin}/ddns/{profile.id}?{profile.ipv4Enabled ? "ipv4=<your_ipv4>" : ""}{profile.ipv4Enabled && profile.ipv6Enabled ? "&" : ""}{profile.ipv6Enabled ? "ipv6=<your_ipv6>" : ""}
         </code>
         <p class="text-gray-600 mt-2">
           Use Basic Auth with username:{" "}
